requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-yasg
drf-spectacular